charms/tensorboards-web-app/requirements-unit.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements-unit.in
#
anyio==3.7.1
    # via httpcore
attrs==23.1.0
    # via jsonschema
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
    #   requests
charmed-kubeflow-chisme==0.2.0
    # via -r requirements.in
charset-normalizer==3.2.0
    # via requests
cosl==0.0.12
    # via -r requirements.in
coverage==7.3.0
    # via -r requirements-unit.in
deepdiff==6.2.1
    # via charmed-kubeflow-chisme
exceptiongroup==1.1.2
    # via
    #   anyio
    #   pytest
h11==0.14.0
    # via httpcore
httpcore==0.17.3
    # via httpx
httpx==0.24.1
    # via lightkube
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
importlib-resources==6.0.1
    # via jsonschema
iniconfig==2.0.0
    # via pytest
jinja2==3.1.2
    # via charmed-kubeflow-chisme
jsonschema==4.17.3
    # via serialized-data-interface
lightkube==0.14.0
    # via
    #   -r requirements.in
    #   charmed-kubeflow-chisme
lightkube-models==1.27.1.4
    # via lightkube
markupsafe==2.1.3
    # via jinja2
oci-image==1.0.0
    # via -r requirements.in
ops==2.14.1
    # via
    #   -r requirements-unit.in
    #   -r requirements.in
    #   charmed-kubeflow-chisme
    #   cosl
    #   serialized-data-interface
ordered-set==4.1.0
    # via deepdiff
packaging==23.1
    # via pytest
pkgutil-resolve-name==1.3.10
    # via jsonschema
pluggy==1.2.0
    # via pytest
pyrsistent==0.19.3
    # via jsonschema
pytest==7.4.0
    # via
    #   -r requirements-unit.in
    #   pytest-lazy-fixture
    #   pytest-mock
pytest-lazy-fixture==0.6.3
    # via -r requirements-unit.in
pytest-mock==3.11.1
    # via -r requirements-unit.in
pyyaml==6.0.1
    # via
    #   -r requirements-unit.in
    #   cosl
    #   lightkube
    #   ops
    #   serialized-data-interface
requests==2.31.0
    # via serialized-data-interface
ruamel-yaml==0.17.32
    # via charmed-kubeflow-chisme
ruamel-yaml-clib==0.2.8
    # via ruamel-yaml
serialized-data-interface==0.7.0
    # via
    #   -r requirements.in
    #   charmed-kubeflow-chisme
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
tenacity==8.2.2
    # via charmed-kubeflow-chisme
tomli==2.0.1
    # via pytest
typing-extensions==4.12.2
    # via cosl
urllib3==2.0.4
    # via requests
websocket-client==1.6.1
    # via ops
zipp==3.16.2
    # via importlib-resources
